refactor desktop lists
