+++ packages/vite-app-ts/src/components/streams/index.tsx
@@ -1,7 +1,5 @@
 import { Empty, Skeleton } from 'antd';
 import { useMemo } from 'react';
-import { truncateAddress } from '~~/helpers';
-import { dateFormat } from '~~/helpers/date-utils';
 import { useVestedTokens } from '~~/hooks';
 import useResponsive, { DisplaySize } from '~~/hooks/use-responsive';
 import StreamListDesktop from './desktop-list';
@@ -34,36 +32,10 @@ const Streams = () => {
     <div className="mt-4">
       {isEmpty ? (
         <Empty image={Empty.PRESENTED_IMAGE_SIMPLE} />
+      ) : isMobile ? (
+        <StreamListMobile />
       ) : (
-        <>
-          {isMobile ? <StreamListMobile /> : <StreamListDesktop />}
-
-          <div className="mb-4 grid grid-cols-5">
-            <p className="uppercase">Address</p>
-            <p className="uppercase">Start/End</p>
-            <p className="uppercase">Streaming</p>
-            <p className="uppercase">Token</p>
-            <p className="uppercase">$</p>
-          </div>
-          <div className="mb-4">
-            {data?.vestedERC20S?.map((vest, index: number) => {
-              const startDate = dateFormat(vest.startTimestamp);
-              const endDate = dateFormat(vest.endTimestamp);
-
-              return (
-                <div className="mb-4 grid grid-cols-5" key={index}>
-                  <p className="mb-0 text-base">{truncateAddress(vest.id)}</p>
-                  <p className="mb-0 text-base">
-                    {startDate} - {endDate}
-                  </p>
-                  <p className="mb-0 text-base">Incoming</p>
-                  <p className="mb-0 text-base">{vest.underlying.symbol}</p>
-                  <p className="mb-0 text-base">$12.20</p>
-                </div>
-              );
-            })}
-          </div>
-        </>
+        <StreamListDesktop list={data?.vestedERC20S} />
       )}
     </div>
   );
